cleaner + added helpers from prev UI
