+++ packages/demos/email-recovery/src/providers/Web3Provider.tsx
@@ -2,11 +2,12 @@ import { WagmiProvider, createConfig } from "wagmi";
 import { baseSepolia } from "wagmi/chains";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ConnectKitProvider, getDefaultConfig } from "connectkit";
+import { PropsWithChildren } from "react";
 
 const connectKitOptions = {
-    walletConnectCTA: 'both',
-    walletConnectName: 'WalletConnect',
-    hideNoWalletCTA: true,
+  walletConnectCTA: "both",
+  walletConnectName: "WalletConnect",
+  hideNoWalletCTA: true,
 };
 
 // TODO Consider https://wagmi.sh/core/api/connectors/safe
@@ -16,19 +17,19 @@ const config = createConfig(
     walletConnectProjectId: import.meta.env.VITE_WALLET_CONNECT_PROJECT_ID,
     appName: "Safe Email Recovery Demo",
     appDescription: "Safe Email Recovery Demo",
-    appUrl: window.location.origin, 
+    appUrl: window.location.origin,
     appIcon: "https://i.imgur.com/46VRTCF.png",
   }),
 );
 
 const queryClient = new QueryClient();
 
-export const Web3Provider = ({ children }) => {
+export const Web3Provider = ({ children }: PropsWithChildren) => {
   return (
     <WagmiProvider config={config}>
       <QueryClientProvider client={queryClient}>
         <ConnectKitProvider options={connectKitOptions}>
-            {children}
+          {children}
         </ConnectKitProvider>
       </QueryClientProvider>
     </WagmiProvider>
